+++ apps/backend/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @stackframe/stack-backend
 
+## 2.8.19
+
+### Patch Changes
+
+- Various changes
+- Updated dependencies
+  - @stackframe/stack-emails@2.8.19
+  - @stackframe/stack-shared@2.8.19
+
+## 2.8.18
+
+### Patch Changes
+
+- Various changes
+- Updated dependencies
+  - @stackframe/stack-shared@2.8.18
+  - @stackframe/stack-emails@2.8.18
+
+## 2.8.17
+
+### Patch Changes
+
+- Various changes
+  - @stackframe/stack-emails@2.8.17
+  - @stackframe/stack-shared@2.8.17
+
 ## 2.8.16
 
 ### Patch Changes