+++ packages/stack-shared/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @stackframe/stack-shared
 
+## 2.8.19
+
+### Patch Changes
+
+- Various changes
+
+## 2.8.18
+
+### Patch Changes
+
+- Various changes
+
+## 2.8.17
+
 ## 2.8.16
 
 ### Patch Changes