+++ packages/stack-ui/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @stackframe/stack-ui
 
+## 2.8.19
+
+### Patch Changes
+
+- Various changes
+- Updated dependencies
+  - @stackframe/stack-shared@2.8.19
+
+## 2.8.18
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.8.18
+
+## 2.8.17
+
+### Patch Changes
+
+- @stackframe/stack-shared@2.8.17
+
 ## 2.8.16
 
 ### Patch Changes